+++ package/lib/src/widgets/meedu_video_player.dart
@@ -144,6 +144,9 @@ class _MeeduVideoPlayerState extends State<MeeduVideoPlayer> {
                     _.customControls =
                         widget.customControls!(context, _, _.responsive);
                   }
+                  if (widget.customCaptionView != null) {
+                    _.customCaptionView = widget.customCaptionView;
+                  }
                   return ExcludeFocus(
                     excluding: _.excludeFocus,
                     child: Stack(
@@ -187,7 +190,7 @@ class _MeeduVideoPlayerState extends State<MeeduVideoPlayer> {
                           responsive: _.responsive,
                           distanceFromBottom:
                               widget.closedCaptionDistanceFromBottom,
-                          customCaptionView: widget.customCaptionView,
+                          customCaptionView: _.customCaptionView,
                         ),
                         if (_.controlsEnabled &&
                             _.controlsStyle == ControlsStyle.primary)